fix unique/secondary/primary key related bug
1. fix unique key insert multi row panic bug
2. show create table information about index and primary key
3. constraint composite primary key is stored in the tae
4. secondary index don't create index table, just satisfy grammer demand

+++ pkg/sql/colexec/deletion/deletion.go
@@ -79,15 +79,17 @@ func Call(_ int, proc *process.Process, arg any) (bool, error) {
 		tmpBat.Clean(proc.Mp())
 
 		for infoNum, info := range p.DeleteCtxs[i].IndexInfos {
-			rel := p.DeleteCtxs[i].IndexTables[infoNum]
-			oldBatch, rowNum := util.BuildUniqueKeyBatch(bat.Vecs[filterColIndex+1:filterColIndex+1+int32(len(p.DeleteCtxs[i].IndexAttrs))], p.DeleteCtxs[i].IndexAttrs, info.Cols, proc)
-			if rowNum != 0 {
-				err := rel.Delete(ctx, oldBatch, info.ColNames[0])
-				if err != nil {
-					return false, err
+			if info.Unique {
+				rel := p.DeleteCtxs[i].IndexTables[infoNum]
+				oldBatch, rowNum := util.BuildUniqueKeyBatch(bat.Vecs[filterColIndex+1:filterColIndex+1+int32(len(p.DeleteCtxs[i].IndexAttrs))], p.DeleteCtxs[i].IndexAttrs, info.Cols, proc)
+				if rowNum != 0 {
+					err := rel.Delete(ctx, oldBatch, info.ColNames[0])
+					if err != nil {
+						return false, err
+					}
 				}
+				oldBatch.Clean(proc.Mp())
 			}
-			oldBatch.Clean(proc.Mp())
 		}
 	}
 	atomic.AddUint64(&p.AffectedRows, affectedRows)
